+++ src/main/java/com/panxoloto/sharepoint/rest/PLGSharepointClient.java
@@ -124,6 +124,16 @@ public interface PLGSharepointClient {
 	JSONObject uploadFile(String folder, Resource resource, JSONObject jsonMetadata) throws Exception;
 
 	/**
+	 * @param folder
+	 * @param resource
+	 * @param fileName
+	 * @param jsonMetadata
+	 * @return
+	 * @throws Exception
+	 */
+	JSONObject uploadFile(String folder, Resource resource, String fileName, JSONObject jsonMetadata) throws Exception;
+	
+	/**
 	 * @param fileServerRelatUrl
 	 * @param jsonMetadata
 	 * @return
@@ -212,4 +222,11 @@ public interface PLGSharepointClient {
 	 * @throws Exception
 	 */
 	Boolean removePermissionToUsers(String folder, List<String> users, Permission permission) throws Exception;
+
+	/**
+	 * @param folderServerRelativeUrl
+	 * @return
+	 * @throws Exception
+	 */
+	JSONObject getFolderFilesByRelativeUrl(String folderServerRelativeUrl) throws Exception;
 }